# docs/api.yaml
openapi: 3.0.3
info:
  title: NutriClinica API
  version: 1.0.0
  description: |
    API do sistema de prontuário nutricional para clínica escola.

    Este contrato é a fonte da verdade. O backend implementa exatamente estes
    endpoints; o frontend consome exatamente estes formatos.
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://nutriclinica-api.onrender.com
    description: Produção
security:
  - bearerAuth: []
...
paths:
...
  /api/pacientes:
    get:
      tags: [Pacientes]
      summary: Lista pacientes visíveis ao usuário
      parameters:
        - { in: query, name: busca, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, default: 0 } }
        - { in: query, name: size, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginaPaciente' }
    post:
      tags: [Pacientes]
      summary: Cadastra paciente
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PacienteRequest' }
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Paciente' }
        '422': { $ref: '#/components/responses/ErroValidacao' }
  /api/pacientes/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Pacientes]
      summary: Detalha paciente
      description: |
        Paciente fora do escopo do usuário responde 404, não 403: confirmar a
        existência de um prontuário alheio já seria vazamento de dado sensível.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Paciente' }
        '404': { $ref: '#/components/responses/NaoEncontrado' }
    put:
      tags: [Pacientes]
      summary: Atualiza paciente
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PacienteRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Paciente' }
  /api/pacientes/{id}/termo:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Pacientes]
      summary: Consulta o termo de consentimento
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TermoConsentimento' }
        '404': { $ref: '#/components/responses/NaoEncontrado' }
    put:
      tags: [Pacientes]
      summary: Registra ou atualiza o termo de consentimento
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TermoConsentimentoRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TermoConsentimento' }
  /api/pacientes/{id}/evolucao:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Pacientes]
      summary: Série histórica de peso, IMC e medidas para gráficos
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PontoEvolucao' }

  # ---------------- ATENDIMENTOS ----------------
  /api/atendimentos:
    get:
      tags: [Atendimentos]
      summary: Lista atendimentos conforme o perfil do usuário
      description: |
        Estagiário vê apenas os próprios. Supervisor vê os dos seus orientados.
        Admin vê todos.
      parameters:
        - { in: query, name: status, schema: { $ref: '#/components/schemas/StatusAtendimento' } }
        - { in: query, name: pacienteId, schema: { type: integer, format: int64 } }
        - { in: query, name: page, schema: { type: integer, default: 0 } }
        - { in: query, name: size, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginaAtendimento' }
    post:
      tags: [Atendimentos]
      summary: Abre um novo atendimento
      description: |
        O estagiário é sempre o usuário autenticado; o corpo não o informa.
        Falha com 409 `SEM_TERMO_CONSENTIMENTO` se o paciente não tiver termo
        LGPD com aceite, e com 422 se o supervisor informado não tiver perfil
        SUPERVISOR ou não orientar o estagiário.
        O número de prontuário é gerado pelo servidor no formato AAAA-NNNNNN,
        sequencial dentro do ano da data da consulta.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pacienteId, supervisorId, dataConsulta]
              properties:
                pacienteId:   { type: integer, format: int64 }
                supervisorId: { type: integer, format: int64 }
                dataConsulta: { type: string, format: date }
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Atendimento' }
        '404': { $ref: '#/components/responses/NaoEncontrado' }
        '409': { $ref: '#/components/responses/Conflito' }
        '422': { $ref: '#/components/responses/ErroValidacao' }
  /api/atendimentos/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Atendimentos]
      summary: Retorna o prontuário completo
      description: |
        Atendimento fora do escopo do usuário responde 404, não 403 — mesma
        razão de /api/pacientes/{id}.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AtendimentoCompleto' }
        '404': { $ref: '#/components/responses/NaoEncontrado' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
  responses:
...
    NaoEncontrado:
      description: Recurso não encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Erro' }
    Conflito:
      description: Operação inválida no estado atual
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Erro' }
    ErroValidacao:
      description: Falha de validação
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Erro' }
  schemas:
    Erro:
      type: object
      properties:
        codigo:    { type: string, example: TRANSICAO_INVALIDA }
        mensagem:  { type: string }
        timestamp: { type: string, format: date-time }
        campos:
          type: array
          items:
            type: object
            properties:
              campo:    { type: string }
              mensagem: { type: string }
    Perfil:
      type: string
      enum: [ESTAGIARIO, SUPERVISOR, ADMIN]
    StatusAtendimento:
      type: string
      enum: [RASCUNHO, EM_REVISAO, APROVADO, DEVOLVIDO_PARA_CORRECAO]
    SecaoProntuario:
      type: string
      enum: [QUEIXA_PRINCIPAL, HISTORIA_CLINICA, MEDICAMENTOS, ANTROPOMETRIA,
             EXAMES, RECORDATORIO, FREQUENCIA_ALIMENTAR, COMPORTAMENTO_ALIMENTAR,
             DIAGNOSTICO, PLANO, METAS]
    Frequencia:
      type: string
      enum: [NUNCA, RARAMENTE, SEMANAL, QUASE_DIARIO, DIARIO]
...
    Usuario:
      type: object
      properties:
        id:     { type: integer, format: int64 }
        nome:   { type: string }
        email:  { type: string, format: email }
        perfil: { $ref: '#/components/schemas/Perfil' }
        ativo:  { type: boolean }
...
    Paciente:
      type: object
      properties:
        id:             { type: integer, format: int64 }
        nome:           { type: string }
        dataNascimento: { type: string, format: date }
        idade:          { type: integer, description: calculada pelo servidor }
        sexo:           { type: string, enum: [FEMININO, MASCULINO, OUTRO, NAO_INFORMADO] }
        racaCor:        { type: string, enum: [BRANCA, PRETA, PARDA, AMARELA, INDIGENA, NAO_INFORMADO] }
        telefone:       { type: string }
        email:          { type: string }
        possuiTermo:    { type: boolean }
    PacienteRequest:
      type: object
      required: [nome, dataNascimento, sexo]
      properties:
        nome:           { type: string, minLength: 3, maxLength: 150 }
        dataNascimento: { type: string, format: date }
        sexo:           { type: string, enum: [FEMININO, MASCULINO, OUTRO, NAO_INFORMADO] }
        racaCor:        { type: string, enum: [BRANCA, PRETA, PARDA, AMARELA, INDIGENA, NAO_INFORMADO] }
        telefone:       { type: string }
        email:          { type: string, format: email }
    TermoConsentimento:
      type: object
      properties:
        aceiteLgpd:          { type: boolean }
        autorizaUsoPesquisa: { type: boolean }
        dataAceite:          { type: string, format: date }
        registradoPor:       { type: string }
        observacoes:         { type: string }
    TermoConsentimentoRequest:
      type: object
      required: [aceiteLgpd, dataAceite]
      properties:
        aceiteLgpd:          { type: boolean }
        autorizaUsoPesquisa: { type: boolean, default: false }
        dataAceite:          { type: string, format: date }
        observacoes:         { type: string }
    Atendimento:
      type: object
      properties:
        id:               { type: integer, format: int64 }
        numeroProntuario: { type: string }
        paciente:         { $ref: '#/components/schemas/Paciente' }
        estagiario:       { $ref: '#/components/schemas/Usuario' }
        supervisor:       { $ref: '#/components/schemas/Usuario' }
        dataConsulta:     { type: string, format: date }
        status:           { $ref: '#/components/schemas/StatusAtendimento' }
        editavel:         { type: boolean, description: se o usuário atual pode editar }
        secoesPreenchidas:
          type: array
          items: { $ref: '#/components/schemas/SecaoProntuario' }
        submetidoEm: { type: string, format: date-time }
        avaliadoEm:  { type: string, format: date-time }
    AtendimentoCompleto:
      allOf:
        - $ref: '#/components/schemas/Atendimento'
        - type: object
          properties:
            queixaPrincipal:       { $ref: '#/components/schemas/QueixaPrincipal' }
            historiaClinica:       { $ref: '#/components/schemas/HistoriaClinica' }
            medicamentos:          { type: array, items: { $ref: '#/components/schemas/Medicamento' } }
            antropometria:         { $ref: '#/components/schemas/Antropometria' }
            exames:                { type: array, items: { $ref: '#/components/schemas/ExameBioquimico' } }
            recordatorio:          { type: array, items: { $ref: '#/components/schemas/Refeicao' } }
            frequenciaAlimentar:   { $ref: '#/components/schemas/FrequenciaAlimentar' }
            comportamentoAlimentar: { $ref: '#/components/schemas/ComportamentoAlimentar' }
            diagnostico:           { $ref: '#/components/schemas/DiagnosticoPes' }
            plano:                 { $ref: '#/components/schemas/PlanoIntervencao' }
            metas:                 { type: array, items: { $ref: '#/components/schemas/Meta' } }
            avaliacao:             { $ref: '#/components/schemas/Avaliacao' }
    QueixaPrincipal:
      type: object
      properties:
        motivo:             { type: string }
        tempoQueixa:        { type: string }
        tratamentoAnterior: { type: string }
        objetivoConsulta:   { type: string }
        observacoes:        { type: string }
    HistoriaClinica:
      type: object
      properties:
        doencasDiagnosticadas:  { type: string }
        alergias:               { type: string }
        intolerancias:          { type: string }
        historicoFamiliar:      { type: string }
        tabagismo:              { type: string, enum: [NUNCA_FUMOU, EX_FUMANTE, FUMANTE] }
        horasSono:              { type: number }
        qualidadeSono:          { type: string, enum: [RUIM, REGULAR, BOA] }
        nivelEstresse:          { type: string, enum: [BAIXO, MODERADO, ALTO] }
        habitoIntestinal:       { type: string, enum: [DIARIO, ALTERNADO, CONSTIPADO, DIARREICO, IRREGULAR] }
        praticaAtividadeFisica: { type: boolean }
        descricaoAtividade:     { type: string }
        observacoes:            { type: string }
    Medicamento:
      type: object
      required: [tipo, nome]
      properties:
        id:         { type: integer, format: int64 }
        tipo:       { type: string, enum: [MEDICAMENTO, SUPLEMENTO] }
        nome:       { type: string }
        dose:       { type: string }
        frequencia: { type: string }
    AntropometriaRequest:
      type: object
      properties:
        pesoKg:            { type: number, minimum: 1,  maximum: 400 }
        alturaCm:          { type: number, minimum: 30, maximum: 250 }
        circCinturaCm:     { type: number }
        circQuadrilCm:     { type: number }
        percentualGordura: { type: number }
        massaMagraKg:      { type: number }
        aferidoEm:         { type: string, format: date }
    Antropometria:
      allOf:
        - $ref: '#/components/schemas/AntropometriaRequest'
        - type: object
          description: Campos abaixo são somente leitura, calculados pelo servidor.
          properties:
            imc:                   { type: number, readOnly: true }
            classificacaoImc:      { type: string, readOnly: true }
            relacaoCinturaQuadril: { type: number, readOnly: true }
            riscoCardiovascular:   { type: string, readOnly: true, enum: [BAIXO, AUMENTADO, MUITO_AUMENTADO] }
            rcqElevada:            { type: boolean, readOnly: true }
    ExameBioquimico:
      type: object
      required: [nomeExame]
      properties:
        id:              { type: integer, format: int64 }
        nomeExame:       { type: string }
        valor:           { type: number }
        unidade:         { type: string }
        dataExame:       { type: string, format: date }
        valorReferencia: { type: string }
    Refeicao:
      type: object
      required: [tipoRefeicao]
      properties:
        id:            { type: integer, format: int64 }
        tipoRefeicao:  { type: string, enum: [DESJEJUM, LANCHE_MANHA, ALMOCO, LANCHE_TARDE, JANTAR, CEIA, OUTRO] }
        horario:       { type: string, example: "07:30" }
        localRefeicao: { type: string }
        ordem:         { type: integer }
        itens:
          type: array
          items:
            type: object
            required: [alimento]
            properties:
              id:            { type: integer, format: int64 }
              alimento:      { type: string }
              quantidade:    { type: string }
              medidaCaseira: { type: string }
              ordem:         { type: integer }
    FrequenciaAlimentar:
      type: object
      properties:
        frutas:           { $ref: '#/components/schemas/Frequencia' }
        verdurasLegumes:  { $ref: '#/components/schemas/Frequencia' }
        ultraprocessados: { $ref: '#/components/schemas/Frequencia' }
        refrigerante:     { $ref: '#/components/schemas/Frequencia' }
        bebidaAlcoolica:  { $ref: '#/components/schemas/Frequencia' }
        cafe:             { $ref: '#/components/schemas/Frequencia' }
        aguaMlDia:        { type: integer, minimum: 0, maximum: 10000 }
        observacoes:      { type: string }
    ComportamentoAlimentar:
      type: object
      properties:
        comeAssistindoTela:   { type: boolean }
        comeRapido:           { type: boolean }
        pulaRefeicoes:        { type: boolean }
        compulsaoAlimentar:   { type: boolean }
        alimentacaoEmocional: { type: boolean }
        restricaoAlimentar:   { type: boolean }
        observacoes:          { type: string }
    DiagnosticoPes:
      type: object
      properties:
        problema:       { type: string }
        etiologia:      { type: string }
        sinaisSintomas: { type: string }
    PlanoIntervencao:
      type: object
      properties:
        objetivos:                  { type: string }
        prescricaoEnergeticaKcal:   { type: integer }
        percCarboidrato:            { type: number }
        percProteina:               { type: number }
        percLipideo:                { type: number }
        estrategiasComportamentais: { type: string }
        educacaoAlimentar:          { type: string }
    Meta:
      type: object
      required: [descricao, prazo]
      properties:
        id:          { type: integer, format: int64 }
        descricao:   { type: string }
        prazo:       { type: string, enum: [CURTO, MEDIO] }
        indicador:   { type: string }
        dataRetorno: { type: string, format: date }
...
    Avaliacao:
      type: object
      properties:
        id:           { type: integer, format: int64 }
        supervisor:   { $ref: '#/components/schemas/Usuario' }
        notaFinal:    { type: number, readOnly: true }
        resultado:    { type: string, enum: [APROVADO, DEVOLVIDO] }
        parecerGeral: { type: string }
        avaliadoEm:   { type: string, format: date-time }
        itens:
          type: array
          items:
            type: object
            properties:
              id:         { type: integer, format: int64 }
              criterio:   { type: string }
              peso:       { type: number }
              nota:       { type: number }
              comentario: { type: string }
...
    PontoEvolucao:
      type: object
      properties:
        dataConsulta:  { type: string, format: date }
        pesoKg:        { type: number }
        imc:           { type: number }
        circCinturaCm: { type: number }
        circQuadrilCm: { type: number }
...
    PaginaPaciente:
      allOf:
        - $ref: '#/components/schemas/Pagina'
        - type: object
          properties:
            content: { type: array, items: { $ref: '#/components/schemas/Paciente' } }
    PaginaAtendimento:
      allOf:
        - $ref: '#/components/schemas/Pagina'
        - type: object
          properties:
            content: { type: array, items: { $ref: '#/components/schemas/Atendimento' } }
    Pagina:
      type: object
      properties:
        totalElements: { type: integer }
        totalPages:    { type: integer }
        number:        { type: integer }
        size:          { type: integer }